% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmaps.data-package.R
\docType{package}
\name{ozmaps.data-package}
\alias{ozmaps.data}
\alias{ozmaps.data-package}
\title{ozmaps.data: Detailed Spatial Data Sets of Australia}
\description{
Provides spatial data sets of Australia. Included are 
 statistical and other regions from the Australian Bureau of Statistics, and state and
 territory boundaries. A compromise of geospatial detail and storage size is used
 to enable the entire country to be provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/ozmaps.data}
  \item Report bugs at \url{https://github.com/mdsumner/ozmaps.data/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
